Change project structure to allow for better parameter estimators implementation
