+++ causal-hub/src/datasets/table/categorical/complete.rs
