+++ causal-hub/src/datasets/table/categorical/mod.rs
@@ -1,5 +1,5 @@
-mod dataset;
-pub use dataset::*;
+mod complete;
+pub use complete::*;
 
 mod evidence;
 pub use evidence::*;
