+++ causal-hub/src/estimators/parameters/bayesian/table/mod.rs
@@ -1,0 +1,1 @@
+mod categorical;
