+++ causal-hub/src/estimators/parameters/bayesian/trajectory/mod.rs
@@ -1,0 +1,1 @@
+mod categorical;
