+++ causal-hub/src/estimators/parameters/maximum_likelihood/table/mod.rs
@@ -1,0 +1,2 @@
+mod categorical;
+mod gaussian;
