+++ causal-hub/src/estimators/parameters/maximum_likelihood/trajectory/mod.rs
@@ -1,0 +1,1 @@
+mod categorical;
